pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.54",
    "langchain-openai>=0.3.14",
    "langsmith>=0.3.32",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "telegram>=0.0.1",
]
